% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object.R
\name{object}
\alias{object}
\title{object}
\usage{
object(type = NULL, access = "public", parent_env = parent.frame())
}
\arguments{
\item{type}{class type, e.g. character, numeric}

\item{access}{class access, public, private, or protected (as in Java)}
}
\description{
object
}
